rename trafficgen -> workspace-traffic
